Change elem_info::implies_or_depends from uint8_t to uint16_t
The implies_or_depends array in elem_info, which is used for reporting
which capabilities a given spird enumeration element implies or depends
on, was too small to hold all possible capability ids, given that the
current largest capability id is 6114. To fix this, the data type was
changed to uint16_t, which also matches how capability ids are stored in
spird files.

+++ spird_builder_main.cpp
@@ -129,7 +129,7 @@ struct elem_info
 
 	uint32_t implies_or_depends_count;
 
-	uint8_t implies_or_depends[127];
+	uint16_t implies_or_depends[127];
 };
 
 struct enum_info
